Create a DuckDB connection to a DATRAS dataset (exchange format)

This function creates a URL for a DATRAS eschange dataset based on the input type and opens the dataset using the duckdbfs::open_dataset function. The type must be one of "HH", "HL", or "CA".

Usage

dr_con_exchange(type = NULL)

Arguments

type

A character string specifying the type of the dataset. Must be one of "HH", "HL", or "CA".

Value

A DuckDB dataset object.

Examples

if (FALSE) { # \dontrun{
  dr_con_exchange("HH")
} # }